requirements.txt
accelerate==1.10.1
aiohappyeyeballs==2.6.1
aiohttp==3.13.0
aiortc==1.9.0
aiosignal==1.4.0
anyio==4.11.0
appnope==0.1.4
asttokens==3.0.0
attrs==25.4.0
audioread==3.0.1
bitsandbytes==0.42.0
black==25.9.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
comm==0.2.3
datasets==4.2.0
debugpy==1.8.17
decorator==5.2.1
dill==0.4.0
evaluate==0.4.6
fastapi==0.115.4
executing==2.2.1
filelock==3.20.0
frozenlist==1.8.0
fsspec==2025.9.0
h11==0.16.0
hf-xet==1.1.10
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.35.3
idna==3.11
ipykernel==7.0.1
ipython==9.6.0
ipython-pygments-lexers==1.1.1
jedi==0.19.2
jinja2==3.1.6
jiwer==4.0.0
joblib==1.5.2
jupyter-client==8.6.3
jupyter-core==5.8.1
lazy-loader==0.4
librosa==0.11.0
llvmlite==0.45.1
markupsafe==3.0.3
matplotlib-inline==0.1.7
more-itertools==10.8.0
mpmath==1.3.0
msgpack==1.1.2
multidict==6.7.0
multiprocess==0.70.16
mypy-extensions==1.1.0
nest-asyncio==1.6.0
networkx==3.5
numba==0.62.1
numpy==2.3.4
openai-whisper==20250625
packaging==25.0
pandas==2.3.3
parso==0.8.5
pathspec==0.12.1
peft==0.17.1
pexpect==4.9.0
pip==25.2
platformdirs==4.5.0
pooch==1.8.2
prompt-toolkit==3.0.52
propcache==0.4.1
psutil==7.1.0
ptyprocess==0.7.0
pure-eval==0.2.3
pyarrow==21.0.0
pydantic==2.9.2
pycparser==2.23
pygments==2.19.2
python-dateutil==2.9.0.post0
pytokens==0.2.0
pytz==2025.2
pyyaml==6.0.3
pyzmq==27.1.0
rapidfuzz==3.14.1
regex==2025.9.18
requests==2.32.5
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.16.2
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
soundfile==0.13.1
soxr==1.0.0
stack-data==0.6.3
sympy==1.14.0
threadpoolctl==3.6.0
tiktoken==0.12.0
tokenizers==0.22.1
torch==2.9.0
torchaudio==2.8.0
torchcodec==0.8.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.57.1
typing-extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn[standard]==0.32.0
wcwidth==0.2.14
xxhash==3.6.0
yarl==1.22.0
